courses/Building RAG Agents with LLMs/course/practice/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
arxiv==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.5
feedparser==6.0.10
ffmpy==0.4.0
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
gradio==4.31.4
gradio_client==0.16.4
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.9
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.7
langchain==0.3.2
langchain-community==0.2.0
langchain-core==0.3.9
langchain-nvidia-ai-endpoints==0.3.0
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langserve==0.2.0
langsmith==0.1.132
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.51.2
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.4
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
ruff==0.6.5
scikit-learn==1.4.2
scipy==1.14.1
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sse-starlette==2.1.0
starlette==0.37.2
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.14.1
yarl==1.11.1
